refactor: migrate TS config to `rewriteRelativeImportExtensions` allowing `.ts` extension
