+++ lib/configs/codeStyle.ts
@@ -2,17 +2,17 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-
 import type { Linter } from 'eslint'
+import type { ConfigOptions } from '../types.d.ts'
+
 import {
 	GLOB_FILES_JAVASCRIPT,
 	GLOB_FILES_TYPESCRIPT,
 	GLOB_FILES_VUE,
-} from '../globs.js'
-import { ConfigOptions } from '../types.js'
+} from '../globs.ts'
 
 import stylistic from '@stylistic/eslint-plugin'
-import l10nPlugin from '../plugins/l10n/index.js'
+import l10nPlugin from '../plugins/l10n/index.ts'
 
 /**
  * Config factory for general code style related rules
