+++ lib/configs/javascript.ts
@@ -2,13 +2,19 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-
 import type { Linter } from 'eslint'
 import type { ConfigOptions } from '../types.d.ts'
+
+import {
+	GLOB_FILES_JAVASCRIPT,
+	GLOB_FILES_TESTING,
+	GLOB_FILES_TYPESCRIPT,
+	GLOB_FILES_VUE,
+} from '../globs.ts'
+
 import eslintRules from '@eslint/js'
-import { GLOB_FILES_JAVASCRIPT, GLOB_FILES_TESTING, GLOB_FILES_TYPESCRIPT, GLOB_FILES_VUE } from '../globs.js'
 import globals from 'globals'
-import nextcloudPlugin from '../plugins/nextcloud/index.js'
+import nextcloudPlugin from '../plugins/nextcloud/index.ts'
 
 /**
  * This config provides the base rules for code quality,
