+++ lib/configs/typescript.ts
@@ -2,12 +2,17 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-
 import type { Linter } from 'eslint'
+import type { ConfigOptions } from '../types.d.ts'
+
+import {
+	GLOB_FILES_TESTING,
+	GLOB_FILES_TYPESCRIPT,
+	GLOB_FILES_VUE,
+} from '../globs.ts'
+import { restrictConfigFiles } from '../utils.ts'
+
 import typescriptPlugin from 'typescript-eslint'
-import { GLOB_FILES_TESTING, GLOB_FILES_TYPESCRIPT, GLOB_FILES_VUE } from '../globs.js'
-import { restrictConfigFiles } from '../utils.js'
-import { ConfigOptions } from '../types.js'
 
 /**
  * Typescript related ESLint rules for Nextcloud
