+++ lib/configs/vue.ts
@@ -2,13 +2,13 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-
 import type { Linter } from 'eslint'
+import type { ConfigOptions } from '../types.d.ts'
+
+import { codeStyle } from './codeStyle.ts'
+import { GLOB_FILES_VUE } from '../globs.ts'
 
 import vuePlugin from 'eslint-plugin-vue'
-import { codeStyle } from './codeStyle.js'
-import { GLOB_FILES_VUE } from '../globs.js'
-import { ConfigOptions } from '../types.js'
 
 const stylisticRules = codeStyle({
 	isLibrary: false,
