+++ lib/plugins/l10n/index.ts
@@ -2,11 +2,12 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-
 import type { ESLint } from 'eslint'
-import { __PACKAGE_VERSION__ } from '../../version.js'
-import RuleEllipsis from './rules/enforce-ellipsis.js'
-import RuleNonBreakingSpace from './rules/non-breaking-space.js'
+
+import { __PACKAGE_VERSION__ } from '../../version.ts'
+
+import RuleEllipsis from './rules/enforce-ellipsis.ts'
+import RuleNonBreakingSpace from './rules/non-breaking-space.ts'
 
 /**
  * ESLint plugin to enforce consistent translations
