+++ lib/plugins/l10n/rules/enforce-ellipsis.test.ts
@@ -4,7 +4,8 @@
  */
 import { RuleTester } from 'eslint'
 import { test } from 'vitest'
-import rule from './enforce-ellipsis.js'
+
+import rule from './enforce-ellipsis.ts'
 
 test('rule: enforce-ellipsis', () => {
 	const ruleTester = new RuleTester()
