+++ lib/plugins/l10n/rules/non-breaking-space.test.ts
@@ -2,11 +2,11 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-
 /* eslint-disable @nextcloud-l10n/non-breaking-space */
 import { RuleTester } from 'eslint'
 import { test } from 'vitest'
-import rule from './non-breaking-space.js'
+
+import rule from './non-breaking-space.ts'
 
 test('rule: non-breaking-space', () => {
 	const ruleTester = new RuleTester()
