+++ lib/plugins/nextcloud/rules/index.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 import type { Rule } from 'eslint'
-import noDeprecations from './no-deprecations.js'
-import noRemovedApis from './no-removed-apis.js'
+import noDeprecations from './no-deprecations.ts'
+import noRemovedApis from './no-removed-apis.ts'
 
 export const rules: Record<string, Rule.RuleModule> = {
 	'no-deprecations': noDeprecations,
