+++ lib/plugins/nextcloud/rules/no-deprecations.ts
@@ -3,7 +3,8 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 import type { Rule } from 'eslint'
-import { createVersionValidator } from '../utils/version-parser.js'
+
+import { createVersionValidator } from '../utils/version-parser.ts'
 
 // ------------------------------------------------------------------------------
 // Rule Definition
