+++ lib/plugins/nextcloud/rules/no-removed-apis.test.ts
@@ -4,7 +4,8 @@
  */
 import { RuleTester } from 'eslint'
 import { describe, test } from 'vitest'
-import rule from './no-removed-apis.js'
+
+import rule from './no-removed-apis.ts'
 
 // ------------------------------------------------------------------------------
 // Tests
