+++ lib/plugins/nextcloud/rules/no-removed-apis.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 import { Rule, Scope } from 'eslint'
-import { createVersionValidator } from '../utils/version-parser.js'
+import { createVersionValidator } from '../utils/version-parser.ts'
 
 // ------------------------------------------------------------------------------
 // Rule Definition
