+++ lib/plugins/packageJson.ts
@@ -4,7 +4,8 @@
  */
 
 import type { ESLint, Rule } from 'eslint'
-import { __PACKAGE_VERSION__ } from '../version.js'
+
+import { __PACKAGE_VERSION__ } from '../version.ts'
 
 import path from 'node:path'
 import sortPackageJson from 'sort-package-json'
