+++ tsconfig.json
@@ -1,8 +1,9 @@
 {
 	"$schema": "https://json.schemastore.org/tsconfig",
 	"compilerOptions": {
-		"esModuleInterop": true,
 		"declaration": true,
+		"esModuleInterop": true,
+		"rewriteRelativeImportExtensions": true,
 		"moduleResolution": "NodeNext",
 		"module": "NodeNext",
 		"target": "ES2022",
